Finalizacion Modulo 4: Seguridad, CORS, Autenticacion y documentacion de entorno local

+++ src/main/java/pe/edu/untels/certificadosdrsu/repositories/ParticipanteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.untels.certificadosdrsu.entities.Participante;
 
 @Repository
-public interface ParticipanteRepository extends JpaRepository<Participante, Integer> {
+public interface ParticipanteRepository extends JpaRepository<Participante, Long> {
     
 }
