+++ src/main/java/pe/edu/untels/certificadosdrsu/servicesinterface/ParticipanteService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface ParticipanteService {
     public List<Participante> list();
     public Participante insert(Participante p);
-    public Optional<Participante> listId(int id);
+    public Optional<Participante> listId(Long id);
     public Participante update(Participante p);
-    public void delete(int id);
+    public void delete(Long id);
 }
